Move popup_handler back to test_menu
